requirements.txt
fastapi
uvicorn
python-jose[cryptography]
pydantic
python-multipart
mysql-connector-python
requests
passlib[bcrypt] 